requirements.txt
flask
google-cloud-logging
google-cloud-pubsub